requirements.txt
# Core framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# LangChain + LangGraph
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.10
langchain-huggingface==0.0.3
langgraph==0.2.28

# Vector store & embeddings
chromadb==0.5.5
sentence-transformers==3.0.1

# Google Gemini SDK
google-generativeai==0.7.2

# Utilities
python-dotenv==1.0.1
pydantic==2.8.2